requirements.txt
# Core LangChain
langchain
langchain-community
langchain-mistralai
torch
transformers 

# FastAPI & Server
fastapi
uvicorn
# fastmcp

# Vector Store & Embeddings
chromadb
sentence-transformers
langchain-chroma

# Data & Utils
requests
pydantic
redis
pandas
python-dotenv

# Graph & State Management
langgraph

# Testing
pytest
pytest-asyncio
httpx

streamlit

huggingface_hub
